requirements.txt
# Dev environment
pip
autopep8

# App
flask
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pandas